requirements.txt
gradio
torch
torchvision
timm
Pillow
huggingface_hub